requirements.txt
pdf2image==1.12.1
Pillow==7.1.1
pytesseract==0.3.3
